requirements.txt
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
Django==3.1.5
django-allauth==0.44.0
django-cors-headers==3.6.0
django-environ==0.4.5
django-rest-auth==0.9.5
djangorestframework==3.12.2
drf-spectacular==0.13.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
jsonschema==3.2.0
oauthlib==3.1.0
Pillow==8.1.0
pycparser==2.20
PyJWT==2.0.1
pyrsistent==0.17.3
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
